# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/neo4j/verify": {
      "get": {
        "operationId": "AppController_verifyNeo4jConnection",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/contracts": {
      "get": {
        "operationId": "ContractsController_getAllContracts",
        "summary": "Get all contracts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns all contract files with their parsed content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContractFileDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/types": {
      "get": {
        "operationId": "ContractsController_getContractTypes",
        "summary": "Get all unique contract types",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns all unique contract types found in the system",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractTypesResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Failed to retrieve contract types"
          }
        },
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/validate": {
      "get": {
        "operationId": "ContractsController_validateContracts",
        "summary": "Validate all contracts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns validation results for all contract files",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/check-if-contract-modified": {
      "get": {
        "operationId": "ContractsController_checkIfContractsModified",
        "summary": "Check if contract files have been modified",
        "description": "Compares current contract file hashes with stored hashes in Neo4j database to detect modifications, additions, and removals",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns information about modified, added, and removed contracts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckModifiedResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Failed to check contract modifications"
          }
        },
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/apply": {
      "post": {
        "operationId": "ContractsController_applyContracts",
        "summary": "Apply contract changes to Neo4j database",
        "description": "Validates all contracts first, then applies them to the Neo4j database if validation passes. Clears existing data before applying.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Contracts successfully applied to database",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Contract validation failed - invalid contracts detected"
          },
          "500": {
            "description": "Failed to apply contracts to database"
          }
        },
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/{module_id}/relations": {
      "get": {
        "operationId": "ContractsController_getModuleRelations",
        "summary": "Get relations for a specific module",
        "description": "Returns outgoing dependencies (modules this module depends on) and incoming dependencies (modules that depend on this module) with the parts they use",
        "parameters": [
          {
            "name": "module_id",
            "required": true,
            "in": "path",
            "description": "The module ID to get relations for",
            "schema": {
              "example": "users-service",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns module relations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleRelationsResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Module not found"
          }
        },
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/search": {
      "get": {
        "operationId": "ContractsController_searchByDescription",
        "summary": "Search modules by description using semantic similarity",
        "description": "Searches for modules using embedding-based semantic similarity. The query description is embedded and compared against stored module embeddings to find the most similar modules.",
        "parameters": [
          {
            "name": "query",
            "required": true,
            "in": "query",
            "description": "The search query text to find similar modules",
            "schema": {
              "example": "user authentication service",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of results to return (default: 10)",
            "schema": {
              "example": 10,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns search results ordered by similarity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchByDescriptionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid search query (empty or missing)"
          },
          "500": {
            "description": "Failed to perform search (embedding service not ready or Neo4j error)"
          }
        },
        "tags": [
          "contracts"
        ]
      }
    }
  },
  "info": {
    "title": "AI Contracts System API",
    "description": "API for managing and validating technical dependency contracts",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "ContractFileDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "description": "The name of the contract file",
            "example": "example-contract.yml"
          },
          "filePath": {
            "type": "string",
            "description": "The full path to the contract file",
            "example": "/contracts/example-contract.yml"
          },
          "content": {
            "type": "object",
            "description": "The parsed contract content"
          },
          "fileHash": {
            "type": "string",
            "description": "SHA256 hash of the contract file content",
            "example": "a3d2f1e8b9c7d6e5f4a3b2c1d0e9f8a7b6c5d4e3f2a1b0c9d8e7f6a5b4c3d2e1"
          }
        },
        "required": [
          "fileName",
          "filePath",
          "content",
          "fileHash"
        ]
      },
      "ContractTypesResponseDto": {
        "type": "object",
        "properties": {
          "types": {
            "description": "Array of unique contract types found in the system",
            "example": [
              "controller",
              "service",
              "component"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "count": {
            "type": "number",
            "description": "Total count of unique contract types",
            "example": 3
          }
        },
        "required": [
          "types",
          "count"
        ]
      },
      "ValidationErrorDto": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "description": "The path to the field with the validation error",
            "example": "dependencies.0.module_id"
          },
          "message": {
            "type": "string",
            "description": "The validation error message",
            "example": "Referenced module \"users-permissions\" does not exist"
          }
        },
        "required": [
          "path",
          "message"
        ]
      },
      "ValidationFileDto": {
        "type": "object",
        "properties": {
          "filePath": {
            "type": "string",
            "description": "The full path to the contract file",
            "example": "/contracts/example-contract.yml"
          },
          "fileName": {
            "type": "string",
            "description": "The name of the contract file",
            "example": "example-contract.yml"
          },
          "valid": {
            "type": "boolean",
            "description": "Whether the contract file is valid",
            "example": true
          },
          "errors": {
            "description": "List of validation errors (only present if valid is false)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorDto"
            }
          }
        },
        "required": [
          "filePath",
          "fileName",
          "valid"
        ]
      },
      "ValidationResponseDto": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean",
            "description": "Whether all contracts are valid",
            "example": true
          },
          "files": {
            "description": "Validation results for each contract file",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationFileDto"
            }
          }
        },
        "required": [
          "valid",
          "files"
        ]
      },
      "ModifiedContractDto": {
        "type": "object",
        "properties": {
          "moduleId": {
            "type": "string",
            "description": "The module ID from the contract",
            "example": "users-get"
          },
          "fileName": {
            "type": "string",
            "description": "The name of the contract file",
            "example": "users-get.yml"
          },
          "filePath": {
            "type": "string",
            "description": "The full path to the contract file",
            "example": "/contracts/users-get.yml"
          },
          "currentHash": {
            "type": "string",
            "description": "The current hash of the contract file",
            "example": "a3d2f1e8b9c7d6e5f4a3b2c1d0e9f8a7b6c5d4e3f2a1b0c9d8e7f6a5b4c3d2e1"
          },
          "storedHash": {
            "type": "string",
            "description": "The stored hash from Neo4j database (null if contract is new)",
            "example": "b4e3d2c1b0a9f8e7d6c5b4a3d2e1f0a9b8c7d6e5f4a3b2c1d0e9f8a7b6c5d4e3",
            "nullable": true
          },
          "status": {
            "type": "string",
            "description": "The status of the contract: 'modified', 'added', or 'removed'",
            "example": "modified",
            "enum": [
              "modified",
              "added",
              "removed"
            ]
          }
        },
        "required": [
          "moduleId",
          "fileName",
          "filePath",
          "currentHash",
          "status"
        ]
      },
      "CheckModifiedResponseDto": {
        "type": "object",
        "properties": {
          "hasChanges": {
            "type": "boolean",
            "description": "Whether any contracts have been modified, added, or removed",
            "example": true
          },
          "totalChanges": {
            "type": "number",
            "description": "Total number of contracts that have changed",
            "example": 3
          },
          "modifiedCount": {
            "type": "number",
            "description": "Number of modified contracts",
            "example": 2
          },
          "addedCount": {
            "type": "number",
            "description": "Number of newly added contracts",
            "example": 1
          },
          "removedCount": {
            "type": "number",
            "description": "Number of removed contracts",
            "example": 0
          },
          "changes": {
            "description": "List of contracts that have been modified, added, or removed",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModifiedContractDto"
            }
          }
        },
        "required": [
          "hasChanges",
          "totalChanges",
          "modifiedCount",
          "addedCount",
          "removedCount",
          "changes"
        ]
      },
      "ApplyResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the operation was successful",
            "example": true
          },
          "modulesProcessed": {
            "type": "number",
            "description": "Number of modules processed",
            "example": 5
          },
          "partsProcessed": {
            "type": "number",
            "description": "Number of parts processed",
            "example": 12
          },
          "message": {
            "type": "string",
            "description": "Success or error message",
            "example": "Successfully applied 5 modules and 12 parts to Neo4j"
          }
        },
        "required": [
          "success",
          "modulesProcessed",
          "partsProcessed",
          "message"
        ]
      },
      "DependencyPartDto": {
        "type": "object",
        "properties": {
          "part_id": {
            "type": "string",
            "description": "Part ID",
            "example": "getUserById"
          },
          "type": {
            "type": "string",
            "description": "Part type",
            "example": "function"
          }
        },
        "required": [
          "part_id",
          "type"
        ]
      },
      "OutgoingDependencyDto": {
        "type": "object",
        "properties": {
          "module_id": {
            "type": "string",
            "description": "Module ID that this module depends on",
            "example": "users-service"
          },
          "parts": {
            "description": "Parts from the dependency module that this module uses",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DependencyPartDto"
            }
          }
        },
        "required": [
          "module_id",
          "parts"
        ]
      },
      "IncomingDependencyDto": {
        "type": "object",
        "properties": {
          "module_id": {
            "type": "string",
            "description": "Module ID that depends on this module",
            "example": "users-controller"
          },
          "parts": {
            "description": "Parts from this module that the dependent module uses",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DependencyPartDto"
            }
          }
        },
        "required": [
          "module_id",
          "parts"
        ]
      },
      "ModuleRelationsResponseDto": {
        "type": "object",
        "properties": {
          "module_id": {
            "type": "string",
            "description": "The module ID being queried",
            "example": "users-service"
          },
          "outgoing_dependencies": {
            "description": "Outgoing dependencies - modules that this module depends on",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OutgoingDependencyDto"
            }
          },
          "incoming_dependencies": {
            "description": "Incoming dependencies - modules that depend on this module and which parts they use",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IncomingDependencyDto"
            }
          }
        },
        "required": [
          "module_id",
          "outgoing_dependencies",
          "incoming_dependencies"
        ]
      },
      "ModuleSearchResultDto": {
        "type": "object",
        "properties": {
          "module_id": {
            "type": "string",
            "description": "The module ID",
            "example": "users-service"
          },
          "type": {
            "type": "string",
            "description": "The module type",
            "example": "service"
          },
          "description": {
            "type": "string",
            "description": "The module description",
            "example": "Service for managing user data and authentication"
          },
          "category": {
            "type": "string",
            "description": "The module category",
            "example": "backend"
          },
          "similarity": {
            "type": "number",
            "description": "Similarity score (0-1, where 1 is most similar)",
            "example": 0.85
          }
        },
        "required": [
          "module_id",
          "type",
          "description",
          "category",
          "similarity"
        ]
      },
      "SearchByDescriptionResponseDto": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string",
            "description": "The search query that was used",
            "example": "authentication service"
          },
          "resultsCount": {
            "type": "number",
            "description": "Number of results returned",
            "example": 5
          },
          "results": {
            "description": "Array of module search results ordered by similarity (highest first)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModuleSearchResultDto"
            }
          }
        },
        "required": [
          "query",
          "resultsCount",
          "results"
        ]
      }
    }
  }
}